+++ src/index.ts
@@ -5,19 +5,25 @@ export interface EventResultMap {
 
 /** 事件回调 */
 export type EventCallback<R = any> = (rs: R) => void
+/** 过滤函数回调 */
+export type FilterCallback<I = any, O = any> = (rs: I) => Promise<O>
+/** 条件性调用回调 */
 export type EventOnceUntilCallback<R = any> = (rs: R) => boolean | undefined
 
 export interface EventFnMap {
   [eventName: string]: EventCallback[]
 }
 export class EventSubscribe<
-  K extends string,
-  M extends Record<K, any>,
-  F extends Record<K, any> = M,
-  R = F[K]
+  M extends EventResultMap,
+  F extends Record<keyof M, any> = M,
+  K extends keyof M = keyof M,
+  R extends F[K] = F[K]
 > {
-  private eventResultMap: Partial<Record<K, R>> = {}
-  private eventFnMap: Partial<Record<K, EventCallback<R>[]>> = {}
+  /** 事件 结果 map */
+  private eventResultMap: Map<K, R> = new Map()
+  /** 事件 filterMap */
+  private eventFilterMap: Map<K, FilterCallback<M[K], R>> = new Map()
+  private eventFnMap: Map<K, EventCallback<R>[]> = new Map()
   /** 事件 key map */
   private eventKeyMap: Map<string, EventCallback<R>> = new Map()
   /** 事件key */
@@ -40,12 +46,18 @@ export class EventSubscribe<
    * @param fnKey: 用于去掉订阅时标识
    * @returns eventKey 订阅标识, 用于 off
    * */
-  on(name: K, done: EventCallback<R>, immediate?: boolean, fnKey?: string) {
+  on<IK extends K, IR = F[IK]>(
+    name: IK,
+    done: EventCallback<IR>,
+    immediate?: boolean,
+    fnKey?: string
+  ) {
     const { eventFnMap, eventResultMap, eventKeyMap } = this
-    if (eventFnMap[name]) {
-      eventFnMap[name]?.push(done)
+    const iEvents = eventFnMap.get(name)
+    if (!iEvents) {
+      eventFnMap.set(name, [done])
     } else {
-      eventFnMap[name] = [done]
+      iEvents.push(done)
     }
 
     if (fnKey) {
@@ -57,8 +69,11 @@ export class EventSubscribe<
     const eventKey = this.formatEventKey(`${name}`, fnKey)
     eventKeyMap.set(eventKey, done)
 
-    if (immediate && name in eventResultMap) {
-      done(eventResultMap[name] as R)
+    if (immediate) {
+      const iResult = eventResultMap.get(name)
+      if (iResult) {
+        done(iResult)
+      }
     }
     return eventKey
   }
@@ -72,7 +87,11 @@ export class EventSubscribe<
    * @param immediate: 立刻执行
    * @returns eventKey 订阅标识, 用于 off
    * */
-  onceUntil(name: K, callback: EventOnceUntilCallback<R>, immediate?: boolean) {
+  onceUntil<IK extends K, IR extends F[IK]>(
+    name: IK,
+    callback: EventOnceUntilCallback<IR>,
+    immediate?: boolean
+  ) {
     const key = this.on(
       name,
       (res) => {
@@ -92,7 +111,7 @@ export class EventSubscribe<
    * @param callback: 回调方法
    * @returns eventKey 订阅标识, 用于 off
    * */
-  once(name: K, callback: EventCallback<R>) {
+  once<IK extends K, IR extends F[IK]>(name: IK, callback: EventCallback<IR>) {
     const key = this.on(
       name,
       (res) => {
@@ -110,9 +129,9 @@ export class EventSubscribe<
    * @param name: 事件名称
    * @param ctx: 订阅时方法 | 订阅标识
    * */
-  off(name: K, ctx: EventCallback<R> | string) {
+  off<IK extends K, IR extends F[IK]>(name: IK, ctx: EventCallback<IR> | string) {
     const { eventFnMap, eventKeyMap } = this
-    const eventFns = eventFnMap[name]
+    const eventFns = eventFnMap.get(name)
     let rFn: EventCallback | undefined
     if (eventFns?.length) {
       if (typeof ctx === 'string') {
@@ -136,38 +155,61 @@ export class EventSubscribe<
    * @param name: 事件名称
    * @param data: 入参数据
    * */
-  trigger(name: K, data: R) {
-    const { eventFnMap, eventResultMap } = this
-    if (eventFnMap[name]) {
-      eventFnMap[name]?.forEach((fn) => {
-        fn(data)
+  async trigger<IK extends K, IR extends M[IK]>(name: IK, data: IR) {
+    const { eventFnMap, eventResultMap, eventFilterMap } = this
+
+    const iFilter = eventFilterMap.get(name)
+    let result: M[K] | R = data
+    if (iFilter) {
+      result = await iFilter(data)
+    }
+
+    const iFns = eventFnMap.get(name)
+    if (iFns) {
+      iFns.forEach((fn) => {
+        fn(result)
       })
     }
-    eventResultMap[name] = data
+    eventResultMap.set(name, result)
   }
 
   /**
    * 事件回放
    * @param name: 事件名称
    * */
-  replay(name: K) {
-    const { eventFnMap, eventResultMap } = this
-    if (eventFnMap[name] && name in eventResultMap) {
-      const lastResult = eventResultMap[name]
-      eventFnMap[name]?.forEach((fn) => {
+  replay<IK extends K>(name: IK) {
+    const { eventFnMap, eventResultMap, eventFilterMap } = this
+    const iFns = eventFnMap.get(name)
+    if (iFns && eventResultMap.has(name)) {
+      const lastResult = eventResultMap.get(name)
+      iFns.forEach((fn) => {
         fn(lastResult as R)
       })
     }
   }
 
+  /**
+   * 添加处理函数
+   * @param name 事件 名称
+   * @param done 过滤方法
+   */
+  async addFilter<IK extends K, IM extends M[IK], IR extends R>(
+    name: IK,
+    done: FilterCallback<IM, IR>
+  ) {
+    const { eventFilterMap } = this
+    eventFilterMap.set(name, done)
+  }
+
+  /** 同 destroy */
   reset() {
     this.destroy()
   }
 
   /** destroy 清空已绑定的事件 */
   destroy() {
-    this.eventResultMap = {}
-    this.eventFnMap = {}
+    this.eventResultMap.clear()
+    this.eventFnMap.clear()
     this.eventKeyMap.clear()
   }
 }